Simplify checkers
This commit moves to an array of classes as the basic data structure and
somewhat inverts the responsibility for knowing which checks should be
used. Instead of the DangerCommitLint class knowing, the checkers
themselves know what their type is.

At the same time, I move the message knowledge into each checker.